feat: unify the header toolbar across single and grid views
Extract the standard (single-view) header into a shared AppToolbar.vue and
use it in both the single and grid (multi-terminal) views, so both show one
identical toolbar instead of the previous icon launcher vs. text/emoji bar.

- AppToolbar reads the global singleton stores (shortcuts / browse /
  accounting / sound) and takes viewMode as a prop.
- Launcher buttons targeting single-view surfaces (Chat / Collections /
  Accounting / favorites) set the global state and emit go-single; the grid
  view switches back so the overlay is shown.
- Active highlighting is gated on the single view, so the grid view only
  highlights the Grid button.
- The grid-only "+ Terminal" action becomes an icon button shown only in the
  grid view.
